requirements.txt
flask
playwright
httpx
pandas
numpy
openai>=1.0.0
python-dotenv
beautifulsoup4